Added tests from MAIN into SANDBOX, then deleted the main suite and removed ctest reference
